Implemented message encryption/decryption in Donkey. Added an Encryptor interface in Donkey and modified the Mirth Server to provide it's encryptor on channel deploy.
Made message exporting work with the encryption.export setting in mirth.properties.

Modified Channel.isConfigurationValid() to verify that each destination has a FilterTransformerExecutor.

Moved importMessage() from the MessageController class to the Channel class in Donkey, since the import logic needs to use the channel's storage settings and encryptor.

Moved FilterTransformerExecutor to be in the same package as the Channel class, so that the Channel can call the protected setEncryptor() method on FilterTransformerExecutor.

Removed the FilterResponse class since it is not used.
Issue: MIRTH-2250