The HL7v2 data type can now accept strict XML data directly if the strict parser is enabled.
The data type plugins no longer require a method to check if the data type is Xml.

The message browser now checks the actual content to determine whether the content is Xml to enable/disable the formatting options, rather than relying on the plugin for that information.

Ran the code formatter.
Issue: MIRTH-2412