Changed response map keys to use d<metaDataId> instead of the connector name, replaced transfer handler prefixes/suffixes with a TransferMode enum, and added a MapTool class in Velocity context which can be used to get map variables with keys that are invalid VTL identifiers.
The destination connector now puts its response in the response map with the key "d#", with "#" being the connector's metadata ID. The metadata ID has been added to the destination connector table of ChannelSetup as well.

The TemplateValueReplacer has been updated with a custom MapTool class that is placed in the context. Whenever loadContextFromMap is called, it places a reference to the map in the MapTool's list. Then during evaluation, calls to MapTool.get() will iterate through each map and return the appropriate object (if it exists in any of the maps).

Drag-and-drop functionality has been changed across the board. Now everything that uses VariableTransferable/MirthVariableList/etc. will pass in a TransferMode enum, which already has the appropriate prefix/suffix internally. In a Velocity context, if the data being transferred constitutes a valid VTL identifier, the behaviour won't change; it'll still be ${varName}. However if the data isn't a valid VTL identifier, the MapTool is used, like ${maps.get('var Name')}. A replacement is done on single quotes as well so that if the key contains a single quote, the maps.get call will use double quotes instead. Also destination drag-and-drop transfers automatically convert to ${d#}, while the full destination name is still preserved in the list.

ConnectorPanel and ConnectorSettingsPanel have been updated with these changes, and specific connectors that override the default (like Database/JavaScript) have been updated as well.

ResponseSettingsPanel has been updated so that the combo box can contain either String objects or Entry objects. This is so that the combo box model can easily keep track of the back-end "d#" key and the front-end full connector name display. When the destination name is changed, and the response variable combo box refreshes its model, it checks on the keys of those entries and re-selects the appropriate element.

Issues: MIRTH-2418, MIRTH-2314, MIRTH-2421, MIRTH-2037