MIRTH-2445, MIRTH-2346, MIRTH-2444: Completely changed how moving/renaming/deleting works in the File Reader.
Instead of "Delete After Read", there is now an "After Processing Action". You can select "None", "Move", or "Delete", depending on what you want to do after a message has been successfully processed through the channel. The Move-to Directory / File Name are still there, and apply to this action.

There are now two different error actions. The first, "Error Reading Action", determines what should happen when the file failed to be read in and processed by the channel.

The other one is "Error in Response Action", which determines what to do when the message was successfully processed, but the status of the selected response is ERROR. In this case, you can choose "After Processing Action", "Move", or "Delete". If "After Processing Action" is chosen (the default), ERROR responses will just cause the After Processing Action to be performed, and not do anything different.

In addition to the error move-to directory, there is also an error move-to file name. Both of these fields apply only to the Error Reading and Error in Response actions.