MIRTH-2459: Paused channels now remain paused after a "Remove All Messages" task runs. The PAUSED and PAUSING states are now also used by the channel itself. Before, the Paused status was inferred by looking at each connector status individually. However now the engine controller just uses the status directly from the channel. The PauseTask and ResumeTask now correctly update the channel status, including setting the state to PAUSING before attempting to stop the source connector. This means that a pausing channel can be halted, just like starting or stopping channels.
The StartTask now also takes a boolean in the constructor, indicating whether or not to start the source connector. DonkeyMessageController then uses the prior state of a channel to determine this. If a channel is being restarted and the prior state was PAUSED or PAUSING, the channel will be started except for the source connector. Recovery will still occur in this case.