A Channel stopped event is now dispatched after a channel is deployed with initial state STOPPED.
Cleaned up ChannelEventType enum. Renamed Deploy to Deployed and Undeploy to Undeployed which is a better description of what actually happens.

Updated the Alert properties map to use private enums as keys instead of creating a static final integer for each key.

Store a reference to channel controller in DefaultAlertWorker to prevent retrieving it each time.
Issue: MIRTH-2891