Added a 'thirdparty' folder to the 'docs' folder in the Server. The folder contains THIRD-PARTY-README.txt which has license information for third party software/libraries that are included in Mirth Connect.
Added several license files that are referred to in THIRD-PARTY-README.txt.

Added third party acknowledgements to the About dialog in the Administrator in order to comply with the license terms of various third party software.

The majority of included third party software/content is covered in this commit, but there remains a handful that still needs proper attribution.
Issue: MIRTH-2958