The order that plugins are loaded can now be set by adding a weight attribute to each serverClass or clientClass. If no weight is specified the plugin class will be given a default weight of 0. Plugins with greater weights will be loaded first. Weights can be negative. If multiple plugins have the same weight, the order they are loaded is random(although usually consistent)
Various plugins are also stored in a LinkedHashMap now rather than normal HashMap. Therefore when they are iterated through they will be in the order they were loaded. This allows us to determine the order for dashboard panel tabs, settings panel tabs, as well as other things. Certain types of plugins are not stored this way because they are typically iterated in alphabetical order. Data types are one example of this. Although they are not stored based on their weight, data type plugins will still be loaded based on them. 
Issue: MIRTH-3034