MIRTH-3393
Implemented a TaskHandler system that allows callbacks when using the new Channel Task system. If tasks are submitted it is now possible to be notified immediately when each task starts, completes, errors, or is cancelled. 

Calling get() on the Futures returned by ChannelUtil tasks will once again throw an exception if the task itself threw an exception. This utilizes the ErrorTaskHandler to retrieve any errors that may have occurred.

Logging exceptions and cancellations is now handled by the LoggingTaskHandler. 

ChannelTask implementations now implement execute(), an abstract method on ChannelTask, instead of call() directly. This allows the task handler callback methods to be injected before and after execute().