New properties have been added to the HL7v2ResponseValidationProperties to provide added functionality to validate the Original Message Control Id.
A boolean variable, validateMessageControlId, will now automatically be set with each new channel. If this value is true, the Original Message Control Id field from the MSH.10 segment will be validated against the received Message Control Id field from the response, MSA.2. All imported channels will have this flag automatically set to false to preserve existing functionality.

An enumeration has been added to store the source of the Original Message Control Id with two values, Destination Encoded and Map Variable. If Destination Encoded is selected, the message control id will be pulled from this field within the response validator, otherwise the Original Message Control Id is taken from a map variable designated by the user.

The final field, originalIdMapVariable, stores the name of the map variable to be used when retrieving and validating the Message Control Id

The HL7v2ResponseValidator now has logic to compare and validate the Original Message Control Id segments if the option is set to true. A method was added to retrieve the original control id based upon the aforementioned options. In the case where the user opted to use a map variable and one was not provided, an exception will be thrown, otherwise the ID is taken from the originating message. If that message is XML, Xpath is used to retrieve the correct segment, otherwise we use string manipulation to retrieve the Original Message Control Id. At last when the IDs are compared, if they are different, the message status is marked as errored with a status message showing the actual vs. expected Ids.
Issue: MIRTH-2163