MIRTH-3205
Fixed problem with the message archiver when an error occurs or the user manually stops the archive process, the MessageWriter.close() method was being skipped, which leaves open the current folder/file being written to. On OS X and Linux the temp folder is automatically wiped out, however in Windows it was not and the user could not delete the file until Mirth Connect was restarted.

Fixed by having MessageWriter.close() get called in a try/finally block.

Added a new finishWrite() method to the MessageWriter interface so that the MessageWriterArchive implementation can have the archive writing logic in this method, rather than having it in close(), which is now always called even on error/halt.

The same problem existed when exporting messages and the same fix has been applied.