MIRTH-2128 The generated script panel was moved to a global scope so that it can be dynamically added or removed on the fly. There now exists a reference to the last selected tab as well as a boolean reference to toggle the generated script panel's visibility. The background color was also made to be consistent with other disabled components, indicating this panel is read only.
The changeListener on the tabbed pane was replaced with a mouseListener as it was interfering with the table's listSelectionListener, and light refactoring was done to bring the variable names more in line with the generated script's title.

When either a javascript or external script type is selected, the generated script panel will now be removed from view. If the selected tab goes from a rule/step to the generated script, the flag is set indicating the panel should automatically switch back to the script view. This will only happen if the script panel is still visible.

Issue: MIRTH-2128