MIRTH-3613
Fixed problem with sending channel alerts when smtp server settings are not configured and an email alert is also configured.

Each alert protocol is wrapped in a try block that catches Exception, so that other alert protocols are still processed if one fails.

Also modified the error message that pops up when smtp settings are not configured so that it makes clear that only email alerts cannot be sent.