MIRTH-3561
Fixed NPE in AlertWorker and NPE in AlertActionPane that occur when the user has an alert with an action protocol that has been uninstalled.

Fixed ConcurrentModificationException in AlertServlet.redactProtocolOptions.

Fixed the behavior of email-sending alert actions (like 'User' and 'Email') so that one email is sent for all these action types (combining recipients), rather than one email for each action type.

Merged the Dispatcher interface into the Protocol interface.