MIRTH-3778
Added a 'updatedUsername' attribute to the LoginStatus class, so that an AuthorizationPlugin can return an updated username after a successful login (for example, an authorization plugin that allows case-insensitive usernames to be entered on login, but must return the case-sensitive matched username that was found in the Mirth Connect database).

Modified all related code that invokes Client.login, so that it obtains the username from the LoginStatus.